Computer Vision type of Rice/deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly
Pillow
opencv-python
scikit-learn==1.4.2
tensorflow==2.15.0